pyproject.toml
[project]
name = "product-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.12",
    "pipecat-ai>=0.0.56",
    "pre-commit>=4.1.0",
    "ruff>=0.9.5",
    "torch>=2.6.0",
    "transformers>=4.48.3",
]

[tools.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
